+++ packages/ledger/src/cosmos/seiLedgerOfflineAminoSigner.ts
@@ -1,4 +1,3 @@
-import Transport from '@ledgerhq/hw-transport-node-hid';
 import {
   AminoSignResponse,
   encodeSecp256k1Signature,
@@ -14,12 +13,12 @@ import { SeiApp } from '@zondax/ledger-sei';
 /**
  *  A signer implementation that uses a Ledger device to sign transactions
  */
-export class LedgerOfflineAminoSigner implements OfflineAminoSigner {
+export class SeiLedgerOfflineAminoSigner implements OfflineAminoSigner {
   private readonly path: string;
   private readonly app: SeiApp;
 
   /**
-   * Creates a new LedgerOfflineAminoSigner
+   * Creates a new SeiLedgerOfflineAminoSigner
    * @param app Ledger Sei app instance
    * @param path hd derivation path (e.g. "m/44'/60'/0'/0/0")
    */
@@ -49,29 +48,3 @@ export class LedgerOfflineAminoSigner implements OfflineAminoSigner {
     };
   }
 }
-
-/**
- * Creates a transport and app instance
- *
- * @returns {Promise<{transport: Transport, app: SeiApp}>} transport and app instances
- */
-export const createTransportAndApp = async () => {
-  const transport = await Transport.create();
-  const app = new SeiApp(transport);
-  return {transport, app};
-};
-
-/**
- * Get the EVM and Cosmos addresses from the Ledger device
- * @param app Ledger Sei app instance
- * @param path hd derivation path (e.g. "m/44'/60'/0'/0/0")
- *
- * @returns {Promise<{evmAddress: string, nativeAddress: string}>} EVM and Cosmos address objects containing
- * address and public key
- */
-export const getAddresses = async (app: SeiApp, path: string) => {
-  const evmAddress = await app.getEVMAddress(path);
-  const nativeAddress = await app.getCosmosAddress(path);
-  return {evmAddress, nativeAddress};
-};
-
