+++ packages/ledger/src/index.ts
@@ -1,1 +1,1 @@
-export * from './wallet';
+export * from './cosmos';
