--- packages/ledger/src/wallet/index.ts
@@ -1,1 +1,0 @@
-export * from "./utils"
